Improve auto trim free thread condition
